refactor: huge change to use javy-plugin 4.0
Move to javy-plugin 4.0. This version requires javy plugins to be
written for the wasip2 target. That requires to use the WebAssembly
component model.

This PR introduces WIT definitions for the javy plugin, plus the import
function that is required by Kubewarden WASI policies.

The WIT definitions make use of the new rich data types provided by the
WIT specification.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>